Make booter support colour console

+++ bare_metal/driver/lowrisc_memory_map.h
@@ -16,7 +16,7 @@ extern volatile uint64_t * get_flash_base();
 extern volatile uint32_t *const plic;
 extern volatile uint64_t *const sd_base;
 extern volatile uint64_t *const sd_bram;
-extern volatile uint8_t *const hid_vga_ptr;
+extern volatile uint16_t *const hid_vga_ptr;
 extern volatile uint64_t *const uart_base;
 extern volatile uint64_t *const eth_base;
